fix detailsFragment and add empty database state

+++ app/src/main/java/com/gap/tinkoffeducation/domain/usecases/GetFeaturesById.kt
@@ -6,7 +6,7 @@ import com.gap.tinkoffeducation.domain.entity.Features
 class GetFeaturesById(
     private val repository: FilmsRepository
 ) {
-    suspend operator fun invoke(id: Long): Features {
+    suspend operator fun invoke(id: Int): Features {
         return repository.getFeaturesById(id)
     }
 }
